remove references to flying saucer
